CRUD for GoalCategory completed

+++ goals/serializers.py
@@ -12,7 +12,7 @@ class GoalCategoryCreateSerializer(serializers.ModelSerializer):
 
 	class Meta:
 		model = GoalCategory
-		read_only_fields = ("id", "created", "updated", "user")
+		read_only_fields = ("id", "created", "updated", "user", "board")
 		fields = "__all__"
 
 
@@ -22,7 +22,7 @@ class GoalCategorySerializer(serializers.ModelSerializer):
 	class Meta:
 		model = GoalCategory
 		fields = "__all__"
-		read_only_fields = ("id", "created", "updated", "user")
+		read_only_fields = ("id", "created", "updated", "user", "board")
 
 
 # Цели
@@ -123,31 +123,32 @@ class BoardSerializer(serializers.ModelSerializer):
 
 	def update(self, instance, validated_data):
 		# owner = validated_data.pop("user")
-		new_participants = validated_data.pop("participants")
-		new_by_id = {part["user"].id: part for part in new_participants}
+		if validated_data.get("participants"):
+			new_participants = validated_data.pop("participants")
+			new_by_id = {part["user"].id: part for part in new_participants}
 
-		old_participants = instance.participants.exclude(user=self.context.get('request').user)
-		with transaction.atomic():
-			for old_participant in old_participants:
-				if old_participant.user_id not in new_by_id:
-					old_participant.delete()
-				else:
-					if (
-							old_participant.role
-							!= new_by_id[old_participant.user_id]["role"]
-					):
-						old_participant.role = new_by_id[old_participant.user_id][
-							"role"
-						]
-						old_participant.save()
-					new_by_id.pop(old_participant.user_id)
-			for new_part in new_by_id.values():
-				BoardParticipant.objects.create(
-					board=instance, user=new_part["user"], role=new_part["role"]
-				)
+			old_participants = instance.participants.exclude(user=self.context.get('request').user)
+			with transaction.atomic():
+				for old_participant in old_participants:
+					if old_participant.user_id not in new_by_id:
+						old_participant.delete()
+					else:
+						if (
+								old_participant.role
+								!= new_by_id[old_participant.user_id]["role"]
+						):
+							old_participant.role = new_by_id[old_participant.user_id][
+								"role"
+							]
+							old_participant.save()
+						new_by_id.pop(old_participant.user_id)
+				for new_part in new_by_id.values():
+					BoardParticipant.objects.create(
+						board=instance, user=new_part["user"], role=new_part["role"]
+					)
 
-			instance.title = validated_data["title"]
-			instance.save()
+		instance.title = validated_data.get("title", instance.title)
+		instance.save()
 
 		return instance
